+++ packages/Peptides/src/utils/misc.ts
@@ -271,6 +271,7 @@ export function modifySelection(selection: type.Selection, clusterOrMonomerPosit
  */
 export function highlightMonomerPosition(monomerPosition: type.SelectionItem, dataFrame: DG.DataFrame,
   monomerPositionStats: MonomerPositionStats): void {
+  if (!dataFrame) return;
   const bitArray = new BitArray(dataFrame.rowCount);
   if (monomerPosition.positionOrClusterType === C.COLUMNS_NAMES.MONOMER) {
     const positionStats = Object.values(monomerPositionStats);
@@ -388,10 +389,13 @@ export function isApplicableDataframe(table: DG.DataFrame, minRows: number = 2):
     wu(table.columns.numerical).toArray().length > 0 && table.rowCount >= minRows;
 }
 
-export function debounce<T extends Array<any>, K>(f: (...args: T) => K, timeout: number = 500): (...args: T) => void {
+export function debounce<T extends Array<any>, K>(f: (...args: T) => Promise<K>, timeout: number = 500,
+): (...args: T) => Promise<K> {
   let timer: NodeJS.Timeout | number | undefined;
-  return (...args: T) => {
-    clearTimeout(timer);
-    timer = setTimeout(() => f(...args), timeout);
+  return async (...args: T) => {
+    return new Promise<K>((resolve) => {
+      clearTimeout(timer);
+      timer = setTimeout(() => resolve(f(...args)), timeout);
+    });
   };
 }